orders/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.4.4
Deprecated==1.2.13
Django==3.2.15
django-environ==0.9.0
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
drf-spectacular==0.24.0
idna==3.3
importlib-resources==5.9.0
inflection==0.5.1
iniconfig==1.1.1
jsonschema==4.16.0
kombu==5.2.4
packaging==21.3
pkgutil_resolve_name==1.3.10
pluggy==1.0.0
prompt-toolkit==3.0.31
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.3
pytest-cov==3.0.0
pytest-django==4.5.2
pytz==2022.2.1
PyYAML==6.0
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.8.1
